+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.0.4
+*2025-01-20*
+
+- Improve NovelAI alpha channel metadata extraction support
+- Detect newer NovelAI images that embed metadata in PNG alpha channel LSB encoding
+
 ## v1.0.3
 *2025-07-28*